requirements.txt
fastapi
uvicorn
uvicorn[standard]
edge_tts
langchain
langchain_community
langchain_huggingface
langchain_core
langchain_google_genai
google-generativeai
sentence-transformers
faiss-cpu
dotenv
pymongo
twilio
websockets
pytest